Worked on DokanY support #5

+++ libqcow/libqcow_file.c
@@ -763,7 +763,7 @@ int libqcow_file_open_file_io_handle(
 	else
 	{
 		internal_file->file_io_handle                   = file_io_handle;
-		internal_file->file_io_handle_opened_in_library = file_io_handle_opened_in_library;
+		internal_file->file_io_handle_opened_in_library = (uint8_t) file_io_handle_opened_in_library;
 	}
 #if defined( HAVE_LIBQCOW_MULTI_THREAD_SUPPORT )
 	if( libcthreads_read_write_lock_release_for_write(
